requirements.txt
python-dotenv==0.18.0
tweepy==3.10.0
